Separate component for inventory report
