+++ apps/web-giddh/src/app/new-inventory/component/report-filters/report-filters.component.scss
@@ -1,0 +1,77 @@
+.stock-mat-salect {
+    border: 1px solid var(--color-secondary-input-bar-border);
+    border-radius: 5px !important;
+    padding: 8px;
+    font-size: 13px;
+    min-width: 250px;
+}
+.transactional-filter-section {
+    .transactional-wise-common-btn {
+        height: 42px;
+        width: 42px;
+        border: 1px solid var(--color-gray);
+        border-radius: 20px;
+        box-shadow: none;
+        background: var(--color-white);
+    }
+
+    .icon-download-circle {
+        font-size: 16px;
+        color: var(--color-gray);
+    }
+}
+.virtual-group-list {
+    height: 193px;
+}
+.select-multiple-chiplist-inventory {
+    font-weight: 400 !important;
+    font-size: 14px !important;
+    width: 82%;
+    .mat-chip.mat-standard-chip {
+        color: var(--color-chips-primary) !important;
+        background-color: var(--color-body-background) !important;
+        border: 1px solid var(--color-primary-login-box-body-text) !important;
+        font-size: 12px !important;
+    }
+}
+.searching-bar-chiplist {
+    width: 83% !important;
+}
+.filter-dropdown-branch {
+    max-width: 250px;
+}
+
+.filter-dropdown-warehouse {
+    max-width: 250px;
+    margin-left: 5px;
+}
+
+.top-main-bar {
+    padding-top: 15px;
+}
+
+@media (max-width: 1440px) {
+    .select-multiple-chiplist-inventory {
+        width: 77%;
+    }
+}
+
+@media (max-width: 1024px) {
+    .select-multiple-chiplist-inventory {
+        width: 57%;
+    }
+
+    .stock-mat-salect {
+        min-width: 187px;
+    }
+}
+
+@media (max-width: 1019px) {
+    .select-multiple-chiplist-inventory {
+        width: 57%;
+    }
+
+    .stock-mat-salect {
+        min-width: 199px;
+    }
+}
